Convert double to integer and update status file.
Review URL: https://chromiumcodereview.appspot.com//10382170

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@7649 260f80e4-7a28-3924-810f-c04153c831b5

+++ lib/compiler/implementation/lib/mockimpl.dart
@@ -251,7 +251,7 @@ class DateImplementation implements Date {
       int minutes = parseIntOrZero(match[5]);
       int seconds = parseIntOrZero(match[6]);
       bool addOneMillisecond = false;
-      int milliseconds = (parseDoubleOrZero(match[7]) * 1000).round();
+      int milliseconds = (parseDoubleOrZero(match[7]) * 1000).round().toInt();
       if (milliseconds == 1000) {
         addOneMillisecond = true;
         milliseconds = 999;
